+++ R/densityfun.R
@@ -155,19 +155,8 @@ function(x,
     if (is.character(width))
       bw <- width
   }
-  if (is.character(bw)) {
-    if (nx < 2)
-      stop("need at least 2 points to select a bandwidth automatically")
-    bw <- switch(tolower(bw), 
-                 nrd0 = stats::bw.nrd0(x), 
-                 nrd = stats::bw.nrd(x),
-                 ucv = stats::bw.ucv(x), 
-                 bcv = stats::bw.bcv(x), 
-                 sj = , 
-                 `sj-ste` = stats::bw.SJ(x, method = "ste"), 
-                 `sj-dpi` = stats::bw.SJ(x, method = "dpi"),
-                 stop("unknown bandwidth rule"))
-  }
+  if (is.character(bw))
+    bw <- bandwidth(x, bw)
   if (!is.finite(bw))
     stop("non-finite 'bw'")
   bw <- adjust * bw